move ServerPingTask in its own file
