+++ launcher/CMakeLists.txt
@@ -922,6 +922,8 @@ SET(LAUNCHER_SOURCES
     ui/pages/instance/McClient.h
     ui/pages/instance/McResolver.cpp
     ui/pages/instance/McResolver.h
+    ui/pages/instance/ServerPingTask.cpp
+    ui/pages/instance/ServerPingTask.h
     
     # GUI - global settings pages
     ui/pages/global/AccountListPage.cpp
